requirements.txt
langchain
faiss-cpu
sentence-transformers
chromadb
stix2
pandas
jinja2
pdfkit
streamlit
python-dotenv
networkx
openai  # for fallback embedding/testing
